v0

+++ mandelbrot.h
@@ -1,5 +1,5 @@
 #include <gtk/gtk.h>
 #include "struct.h"
 
-void mandelbrot(cairo_t* cr, point* co, size_t a, MandelbrotState* b,
+void mandelbrot(cairo_t* cr, point* co, int a, MandelbrotState* b,
         AppSettings* c, int w, int h);
